+++ packages/provider/src/transports/wallet-request-handler.ts
@@ -10,6 +10,7 @@ import {
   ConnectDetails,
   PromptConnectDetails,
   WalletSession,
+  OpenWalletIntent,
   ErrSignedInRequired,
   ProviderEventTypes,
   TypedEventEmitter
@@ -31,6 +32,8 @@ import { isSignedTransactionBundle, TransactionRequest } from '@0xsequence/trans
 import { signAuthorization, AuthorizationOptions } from '@0xsequence/auth'
 import { logger, TypedData } from '@0xsequence/utils'
 
+const SIGNER_READY_TIMEOUT = 10000
+
 export interface WalletSignInOptions {
   connect?: boolean
   mainnetNetworks?: Networks
@@ -39,11 +42,17 @@ export interface WalletSignInOptions {
 }
 
 export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, ProviderMessageRequestHandler {
-  private account: Account | null
+  // signer interface of the wallet. A null value means there is no signer (ie. user not signed in). An undefined
+  // value means the signer state is unknown, usually meaning the wallet app is booting up and initializing. Of course
+  // a Signer value is the actually interface to a signed-in account
+  private account: Account | null | undefined
+  private signerReadyCallbacks: Array<() => void> = []
+
   private prompter: WalletUserPrompter | null
   private mainnetNetworks: NetworkConfig[]
   private testnetNetworks: NetworkConfig[]
 
+  private _openIntent?: OpenWalletIntent
   private _connectOptions?: ConnectOptions
   private _defaultNetworkId?: string | number
   private _chainId?: number
@@ -51,19 +60,19 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
   private events: TypedEventEmitter<ProviderEventTypes> = new EventEmitter() as TypedEventEmitter<ProviderEventTypes>
 
   constructor(
-    signer: Account | null,
+    account: Account | null | undefined,
     prompter: WalletUserPrompter | null,
     mainnetNetworks: Networks,
     testnetNetworks: Networks = []
   ) {
-    this.account = signer
+    this.account = account
     this.prompter = prompter
     this.mainnetNetworks = mainnetNetworks
     this.testnetNetworks = testnetNetworks
   }
 
-  async signIn(signer: Account | null, options: WalletSignInOptions = {}) {
-    this.account = signer
+  async signIn(account: Account | null, options: WalletSignInOptions = {}) {
+    this.setAccount(account)
 
     const { connect, mainnetNetworks, testnetNetworks, defaultNetworkId } = options
 
@@ -105,6 +114,33 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
         this.notifyClose()
       }
     }
+  }
+
+  signOut() {
+    // signed out state
+    this.setAccount(null)
+  }
+
+  accountReset() {
+    // resetting signer puts the wallet in an uninitialized state, which requires the app to
+    // re-initiatize and set the signer either as "null" (ie. no signer) or "Signer" (ie. signed in).
+    this.account = undefined
+  }
+
+  signerReady(timeout: number = SIGNER_READY_TIMEOUT): Promise<void> {
+    return new Promise((resolve, reject) => {
+      if (this.account !== undefined) {
+        resolve()
+      } else {
+        setTimeout(() => {
+          if (this.account === undefined) {
+            this.signerReadyCallbacks = []
+            reject(`signerReady timed out`)
+          }
+        }, timeout)
+        this.signerReadyCallbacks.push(resolve)
+      }
+    })
   }
 
   async connect(options?: ConnectOptions): Promise<ConnectDetails> {
@@ -205,6 +241,8 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
       id: request.id!,
       result: null
     }
+
+    await this.getAccount()
 
     try {
       // only allow public json rpc method to the provider when user is not logged in, aka signer is not set
@@ -581,6 +619,14 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
     }
   }
 
+  get openIntent(): OpenWalletIntent | undefined {
+    return this._openIntent
+  }
+
+  setOpenIntent(intent: OpenWalletIntent | undefined) {
+    this._openIntent = intent
+  }
+
   get connectOptions(): ConnectOptions | undefined {
     return this._connectOptions
   }
@@ -676,16 +722,28 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
     this.events.emit('close', error)
   }
 
-  isSignedIn(): boolean {
+  isSignedIn = async (): Promise<boolean> => {
+    await this.signerReady()
     return !!this.account
   }
 
-  getSigner(): Signer | null {
+  getAccount = async (): Promise<Account | null> => {
+    await this.signerReady()
+    if (this.account === undefined) {
+      throw new Error('signerReady failed resolve')
+    }
     return this.account
   }
 
-  setAccount(account: Account | null) {
+  setAccount(account: Account | null | undefined) {
     this.account = account
+
+    if (account !== undefined) {
+      for (let i = 0; i < this.signerReadyCallbacks.length; i++) {
+        this.signerReadyCallbacks[i]()
+      }
+      this.signerReadyCallbacks = []
+    }
   }
 
   private async handleConfirmWalletDeployPrompt(